Redone AuditLogChange to move generic types into getters instead of the class Moved ActionType and TargetType into core package
